+++ torchgeo/datasets/inaturalist.py
@@ -91,7 +91,7 @@ class INaturalist(GeoDataset):
                 f'query: {query} not found in index with bounds: {self.bounds}'
             )
 
-        keypoints = torch.tensor(index.get_coordinates().values)
+        keypoints = torch.tensor(index.get_coordinates().values, dtype=torch.float32)
         sample = {'crs': self.crs, 'bounds': query, 'keypoints': keypoints}
 
         return sample
